% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\docType{data}
\name{cc}
\alias{cc}
\title{IPEDS (Integrated Postsecondary Education Data System) data, 2017}
\format{A data frame with 6,440 rows and 60 columns. Each row corresponds to an institution.

\describe{
\item{unitid}{(character) unique ID number assigned to postsecondary institutions surveyed through IPEDS}
\item{instnm}{(character) institution name}
\item{state}{(factor) US Postal Service state abbreviation}
\item{addr}{(character) street address or post office box}
\item{zip}{(factor) 5-digit ZIP code}
\item{gentele}{(character) ten-digit telephone number}
\item{webaddr}{(character) Internet website address}
\item{longitud}{(double) longitude of location}
\item{latitude}{(double) latitude of location}
\item{urbanicity}{(factor) whether the location of the institution is \code{rural} or in a \code{town}, \code{city}, or \code{suburb}}
\item{degree_level}{(factor) highest level of degree offered; \code{1} = four or more years, \code{2} = at least two but less than four years, \code{3} = less than two years}
}}
\source{
\url{https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx}
}
\usage{
cc
}
\description{
A dataset formed by extracting information from the surveys conducted by the Institute of Education Sciences in 2017 (link provided as a source). These surveys are conducted annually by NCES () to gather information from every postsecondary institution that participates in federal student financial aid programs.
}
\details{
This dataset represents a population of both degree-granting and non-degree-granting active post-secondary institutions of all levels and sectors. It includes public, private for-profit, and private not-for-profit institutions. The vast majority of institutions in the dataset are Title IV institutions (which participate in federal financial aid programs), although some non-Title IV institutions participate as well.
}
\keyword{datasets}
