% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_help.R
\name{inference_help}
\alias{inference_help}
\title{Inference Help}
\usage{
inference_help()
}
\arguments{
\item{N/A}{no parameters}
}
\value{
Console text teaching people how to filter
}
\description{
This function introduced the dataset and gives filtering recommendations for categorical and continuous variables in the \code{thegeneralizer} package.
}
\examples{
Run inference_help() to see what it does
inference_help()
}
