% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_table.R
\name{mean_table}
\alias{mean_table}
\title{Calculate and visualize the means of stratifying variables across strata.}
\usage{
mean_table(solution, plots = TRUE)
}
\arguments{
\item{solution}{object storing the output of \code{stratifier}}

\item{plots}{defaults to \code{TRUE}; whether or not to create plots}
}
\value{
a table of the mean values of selected stratifying variables across strata
}
\description{
This function does stuff 'n' thangs.
}
\details{
This paragraph talks more about the stuff 'n' thangs.
}
\examples{
\dontrun{
# To view a table of means along with graphs:
mean_table(solution)
}
}
\seealso{
\url{http://thegeneralizer.org/}, also add other resources
}
