% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment_list.R
\name{recruitment_list}
\alias{recruitment_list}
\title{Generate recruitment lists with added id variables based on universe dataset originated from}
\usage{
recruitment_list(solution, custom_id)
}
\arguments{
\item{solution}{object storing the output of \code{stratifier}, typically \code{generalizer_output}}

\item{custom_id}{a dataframe containing the ID Variable and additional id variables to be added to recruitment lists}
}
\value{
Saves recruitment lists in project directory.
}
\description{
Generate recruitment lists with added id variables based on universe dataset originated from
}
