% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schools_table.R
\name{schools_table}
\alias{schools_table}
\title{Generate recruitment lists to achieve a target sample size.}
\usage{
schools_table(solution, number, sample = NULL)
}
\arguments{
\item{solution}{sdfdf}

\item{number}{sdfsdfsd}

\item{sample}{Defaults to \code{NULL}.}
}
\value{
thangs
}
\description{
The input to this function must be the output of the \code{stratifer} function,
along with the sample size that the researcher aims to recruit (typically the
desired number of schools).
}
\details{
As output, this function prints a table with columns equal to the number of
strata specified in \code{stratifier()} and three rows. The first row presents
the number of schools in the inference population that fall in each stratum. The
second row presents the proportion of the population in each stratum. Finally, the
third row contains the number of schools per stratum that the researcher should
aim to recruit in order to ensure that the proportions in their sample mirror
those in the population.
}
\examples{
\dontrun{
# For an experiment where the researcher wants to recruit 100 schools:
schools_table(solution, number = 100)

# For an experiment that has been conducted; the researcher wanted
# to recruit 100 schools and has uploaded their school IDs:
schools_table(solution, number = 100, sample = idvars)
}
}
\seealso{
\url{http://thegeneralizer.org/}, also add other resources
}
