% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_functions.R
\name{set_percentile_limits}
\alias{set_percentile_limits}
\title{Return rows which values lie within a specified percentile interval.}
\usage{
set_percentile_limits(df, var, min_pct, max_pct)
}
\arguments{
\item{df}{the data frame to transform}

\item{var}{the column corresponding to selected variable}

\item{min_quant}{lower boundary percentile}

\item{max_quant}{upper boundary percentile}
}
\value{
A filtered dataframe, containing only rows meeting specified condition.
}
\description{
Use \code{set_percentile_limits} to find rows of a dataframe whose values lies within a specified percentile interval of selected variable.
}
\details{
This function is compatible with dplyr. Use this to set limits on continuous variables within your data.
}
\examples{
# Limiting the inference population to the middle 50\% of schools in terms of their
# percentage of female students (using the post-secondary data).
set_limits(df = ipeds, var = PCT_Female, min_pct = 25, max_pct = 75)
}
