% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifier.R
\name{stratifier}
\alias{stratifier}
\title{Stratify a population data frame.}
\usage{
stratifier(x, metric = "gower")
}
\arguments{
\item{metric}{A character string specifying the metric to be used in
calculating distances for observations. The default is \code{gower}; other
options are \code{euclidean} and \code{manhattan}.}

\item{data}{A data frame containing information about schools in your
inference population. Must contain a column of ID numbers identifying each
school, along with columns of stratifying variables.}
}
\value{
A list. The first element contains the raw results of stratifying the
 data into the user-selected number of clusters. The second element contains
 several lists, each ranking the observations within clusters.
}
\description{
This function is designed to receive a data frame containing information about
an inference population.
}
\details{
...

To complete the generalization process, the user should run this function on
their desired inference population data frame and save the output to an object.
The resulting object, referred to as \code{solution} in the package documentation,
should then be provided as input to the \code{mean_table()} and \code{schools_table()}
functions.
}
\examples{
\dontrun{
stratifier(x)
}
}
\seealso{
\url{http://thegeneralizer.org/}, also add other resources
}
