% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_table.R
\name{mean_table}
\alias{mean_table}
\title{Calculate and visualize the means of stratifying variables across strata.}
\usage{
mean_table(solution, plots = TRUE, store = FALSE)
}
\arguments{
\item{solution}{object storing the output of \code{stratifier}; usually called \code{generalizer_output}}

\item{plots}{defaults to \code{TRUE}; whether or not to create plots}

\item{store}{defaults to \code{FALSE}; whether to export table and plots to files}
}
\value{
a table of the mean values of selected stratifying variables across strata
}
\description{
This function is designed to be run after using \code{stratifier} on the output that is produced. It produces a table of the population means and standard deviations of the stratification variables selected in each cluster. It also presents some visualizations of the population cluster information, which can help users describe the clusters.
}
\details{
The function by default does not store this information, but users can save the plots manually, or specify the argument \code{store} to export the table and plots to files in their current working directory.
}
\examples{
\dontrun{
# To view a table of means along with graphs:
mean_table(solution)
}
}
\seealso{
\url{http://thegeneralizer.org/}, also add other resources
}
