% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifier.R, R/stratifier_exp.R,
%   R/stratifier_v2.R
\name{stratifier}
\alias{stratifier}
\title{Stratify a population data frame.}
\usage{
stratifier(x, metric = "gower")

stratifier(x, metric = "gower")

stratifier(x, metric = "gower")
}
\arguments{
\item{metric}{A character string specifying the metric to be used in
calculating distances for observations. The default is \code{gower}; other
options are \code{euclidean} and \code{manhattan}.}

\item{data}{A data frame containing information about schools in your
inference population. Must contain a column of ID numbers identifying each
school, along with columns of stratifying variables.}

\item{data}{A data frame containing information about schools in your
inference population. Must contain only a column of ID numbers identifying each
school, and columns of relevant stratifying variables.}

\item{idnum}{A character string specifying the name of the ID column identifying each school/observation.}

\item{metric}{A character string specifying the metric to be used in
calculating distances for observations. The default is \code{gower}; other
options are \code{euclidean} and \code{manhattan}.}

\item{clusters}{A number specifying the number of clusters for stratification.}

\item{autosave}{If \code{TRUE}, results of stratification will be saved into global environment as '\code{generalizer_output}'}

\item{data}{A data frame containing information about schools in your
inference population. Must contain a column of ID numbers identifying each
school, along with columns of stratifying variables.}

\item{metric}{A character string specifying the metric to be used in
calculating distances for observations. The default is \code{gower}; other
options are \code{euclidean} and \code{manhattan}.}
}
\value{
A list. The first element contains the raw results of stratifying the
 data into the user-selected number of clusters. The second element contains
 several lists, each ranking the observations within clusters.

A list. The first element contains the raw results of stratifying the
 data into the user-selected number of clusters. The second element contains
 several lists, each ranking the observations within clusters.

A list. The first element contains the raw results of stratifying the
 data into the user-selected number of clusters. The second element contains
 several lists, each ranking the observations within clusters.
}
\description{
This function is designed to receive a data frame containing information about
an inference population.

An expedited version of stratifier. This function is designed to receive a data frame containing information about
an inference population, and containing only stratification variables

This function is designed to receive a data frame containing information about
an inference population.
}
\details{
...

To complete the generalization process, the user should run this function on
their desired inference population data frame and save the output to an object.
The resulting object, referred to as \code{solution} in the package documentation,
should then be provided as input to the \code{mean_table()} and \code{schools_table()}
functions.

...

To complete the generalization process, the user should run this function on
their desired inference population data frame and save the output to an object.
The resulting object, referred to as \code{solution} in the package documentation,
should then be provided as input to the \code{mean_table()} and \code{schools_table()}
functions.
}
\examples{
\dontrun{
stratifier(x)
}
\dontrun{
stratifier(x)
}
}
\seealso{
\url{http://thegeneralizer.org/}, also add other resources

\url{http://thegeneralizer.org/}, also add other resources

\url{http://thegeneralizer.org/}, also add other resources
}
